+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ lib_xud change log
 UNRELEASED
 ----------
 
-  * CHANGE: Optimised port timings for XS3A to use non-inverted clock
+  * ADDED:    Option XUD_SUSPEND_PHY to place USB phy in low power mode during
+    USB suspend
+  * CHANGE:   Optimised port timings for XS3A to use non-inverted clock
 
 2.4.0
 -----